microsoft_defender_endpoint: add support for oauth endpoint params (#15667)
Add support for the oauth_endpoint_params configuration
parameter for all available data streams.

Log data stream still works under httpjson so the option
has been added under data stream level along with all
the OAuth2 options for this data stream.

For the another data streams, as they work under the
CEL input, it has been added at input level so adding
any value to this option will affect all data streams that
rely on CEL (machine, machine_action, and vulnerability).

Finally, the auth logic for the vulnerability data stream is
implemented in the CEL program instead of delegate in
the CEL auth options for the input.
Therefore, the oauth endpoint params in this case are added
manually in the program as well.